Generando imagen escalada Obteniendo error Chi2 mediante reduccion Agregada funcion de cuMyPeakArray para adjuntar más Peaks (Copiando a memoria de Host... Reescribir más adelante) Al parecer todo va bien.
git-svn-id: http://subversion.assembla.com/svn/particle-tracking-2/branches/Chi2HDLib_Cuda@136 ae994a31-0241-454f-aaba-ab632ab2c510

+++ src/Headers/Chi2Libcu.h
@@ -44,7 +44,7 @@ public:
 	/**
 	 * Calcula la diferencia con la Imagen Chi2 y la Imagen normal
 	 */
-	static int computeDifference(cuMyMatrix *img,cuMyMatrix *grid_x, cuMyMatrix *grid_y, float d, float w, cuMyMatrix *diffout);
+	static float computeDifference(cuMyMatrix *img,cuMyMatrix *grid_x, cuMyMatrix *grid_y, float d, float w, cuMyMatrix *diffout);
 };
 
 #endif /* CHI2LIBCU_H_ */
